+++ packages/backend/src/routers/mcpRouter.ts
@@ -6,14 +6,18 @@ import express from 'express';
 import { IncomingMessage } from 'http';
 
 import {
+    ApiKeyAccount,
     ForbiddenError,
     getErrorMessage,
     LightdashError,
+    LightdashSessionUser,
     MissingConfigError,
     NotImplementedError,
     OauthAccount,
+    OauthAuth,
 } from '@lightdash/common';
 import {
+    allowApiKeyAuthentication,
     allowOauthAuthentication,
     isAuthenticated,
 } from '../controllers/authentication';
@@ -61,6 +65,7 @@ const returnHeaderIfUnauthenticated = (
 mcpRouter.all(
     '/',
     allowOauthAuthentication,
+    allowApiKeyAuthentication,
     returnHeaderIfUnauthenticated,
     async (req, res) => {
         try {
@@ -123,6 +128,20 @@ mcpRouter.all(
                         token: oauthAuth.authentication.token,
                         clientId: oauthAuth.authentication.clientId,
                         scopes: oauthAuth.authentication.scopes,
+                        extra,
+                    };
+                }
+
+                if (req.user && req.account?.isPatUser()) {
+                    const apiKeyAuth = req.account as ApiKeyAccount;
+                    const extra: ExtraContext = {
+                        user: req.user,
+                        account: apiKeyAuth,
+                    };
+                    authReq.auth = {
+                        token: apiKeyAuth.authentication.source,
+                        clientId: 'API key', // hardcoded client and scopes for PAT authentication
+                        scopes: ['mcp:read', 'mcp:write'],
                         extra,
                     };
                 }