+++ packages/common/src/ee/commercialFeatureFlags.ts
@@ -2,7 +2,6 @@ export enum CommercialFeatureFlags {
     Embedding = 'embedding',
     Scim = 'scim-token-management',
     AiCopilot = 'ai-copilot',
-    AiAgent = 'ai-agent',
     ServiceAccounts = 'service-accounts',
     OrganizationWarehouseCredentials = 'organization-warehouse-credentials',
 }